@@ -1,6 +1,6 @@
 {
   "name": "amis-core",
-  "version": "2.0.2",
+  "version": "2.1.0",
   "description": "amis-core",
   "main": "lib/index.js",
   "module": "esm/index.js",
@@ -21,8 +21,8 @@
     "jest": "^28.1.0",
     "jest-environment-jsdom": "^28.1.0",
     "moment-timezone": "^0.5.34",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rimraf": "^3.0.2",
     "rollup": "^2.73.0",
     "rollup-plugin-auto-external": "^2.0.0",
@@ -43,7 +43,7 @@
     "esm"
   ],
   "dependencies": {
-    "amis-formula": "^2.0.2",
+    "amis-formula": "^2.1.0",
     "classnames": "2.3.1",
     "file-saver": "^2.0.2",
     "hoist-non-react-statics": "^3.3.2",
@@ -52,7 +52,7 @@
     "mobx": "^4.5.0",
     "mobx-react": "^6.3.1",
     "mobx-state-tree": "^3.17.3",
-    "moment": "^2.19.3",
+    "moment": "^2.19.4",
     "papaparse": "^5.3.0",
     "qs": "6.9.7",
     "react-json-view": "1.21.3",
