@@ -3,7 +3,7 @@
   "main": "lib/index.js",
   "module": "esm/index.js",
   "types": "lib/index.d.ts",
-  "version": "2.0.2",
+  "version": "2.1.0",
   "description": "",
   "scripts": {
     "build": "npm run clean-dist && NODE_ENV=production rollup -c ",
@@ -33,12 +33,12 @@
     }
   },
   "dependencies": {
-    "amis-core": "^2.0.2",
-    "amis-formula": "^2.0.2",
+    "amis-core": "^2.1.0",
+    "amis-formula": "^2.1.0",
     "classnames": "2.3.1",
     "codemirror": "^5.63.0",
     "downshift": "6.1.7",
-    "echarts": "5.3.1",
+    "echarts": "5.3.3",
     "froala-editor": "3.1.1",
     "hoist-non-react-statics": "^3.3.2",
     "jsbarcode": "^3.11.5",
@@ -50,7 +50,7 @@
     "mobx": "^4.5.0",
     "mobx-react": "^6.3.1",
     "mobx-state-tree": "^3.17.3",
-    "moment": "^2.19.3",
+    "moment": "^2.19.4",
     "monaco-editor": "0.30.1",
     "prop-types": "^15.6.1",
     "rc-input-number": "^7.3.4",
@@ -63,7 +63,7 @@
     "react-transition-group": "4.4.2",
     "react-visibility-sensor": "5.1.1",
     "sortablejs": "1.14.0",
-    "tinymce": "^5.10.3",
+    "tinymce": "^6.1.2",
     "tslib": "^2.3.1",
     "uncontrollable": "7.2.1"
   },
@@ -83,15 +83,15 @@
     "jest-environment-jsdom": "^28.1.0",
     "moment-timezone": "^0.5.34",
     "postcss-import": "^14.1.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rimraf": "^3.0.2",
     "rollup": "^2.73.0",
     "rollup-plugin-auto-external": "^2.0.0",
     "rollup-plugin-license": "^2.7.0",
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-scss": "^3.0.0",
-    "sass": "^1.52.1",
+    "sass": "^1.54.0",
     "ts-jest": "^28.0.3",
     "typescript": "^4.6.4"
   },
