@@ -1,6 +1,6 @@
 {
   "name": "amis",
-  "version": "2.0.2",
+  "version": "2.1.0",
   "description": "一种MIS页面生成工具",
   "main": "lib/index.js",
   "module": "esm/index.js",
@@ -40,14 +40,14 @@
     ]
   },
   "dependencies": {
-    "amis-core": "^2.0.2",
-    "amis-ui": "^2.0.2",
+    "amis-core": "^2.1.0",
+    "amis-ui": "^2.1.0",
     "ansi-to-react": "^6.1.6",
     "attr-accept": "2.2.2",
     "blueimp-canvastoblob": "2.1.0",
     "classnames": "2.3.1",
     "downshift": "6.1.7",
-    "echarts": "5.3.1",
+    "echarts": "5.3.3",
     "echarts-stat": "^1.2.0",
     "exceljs": "^4.3.0",
     "file-saver": "^2.0.2",
@@ -62,11 +62,11 @@
     "mobx": "^4.5.0",
     "mobx-react": "^6.3.1",
     "mobx-state-tree": "^3.17.3",
-    "moment": "^2.19.3",
+    "moment": "^2.19.4",
     "monaco-editor": "0.30.1",
     "mpegts.js": "^1.6.10",
     "prop-types": "^15.6.1",
-    "qrcode.react": "^3.0.0",
+    "qrcode.react": "^3.1.0",
     "rc-overflow": "^1.2.4",
     "react-cropper": "^2.1.8",
     "react-dropzone": "^11.4.2",
@@ -111,7 +111,7 @@
     "@types/tinymce": "^4.5.24",
     "axios": "0.25.0",
     "bce-sdk-js": "^0.2.9",
-    "concurrently": "^7.0.0",
+    "concurrently": "^7.3.0",
     "copy-to-clipboard": "3.3.1",
     "core-js": "^3.21.0",
     "css": "3.0.0",
@@ -135,8 +135,8 @@
     "prettier": "^2.6.1",
     "pretty-quick": "^3.1.1",
     "prismjs": "^1.25.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router": "5.2.1",
     "react-router-dom": "5.3.0",
     "react-test-renderer": "^18.0.0",
