@@ -45,6 +45,7 @@ import ColmunsIcon from '../icons/columns.svg';
 import CalendarIcon from '../icons/calendar.svg';
 import ClockIcon from '../icons/clock.svg';
 import TreeDownIcon from '../icons/tree-down.svg';
+import CheckCircle from '../icons/check-circle.svg';
 import CloudUploadIcon from '../icons/cloud-upload.svg';
 import ImageIcon from '../icons/image.svg';
 import RefreshIcon from '../icons/refresh.svg';
@@ -95,9 +96,16 @@ import Role from '../icons/role.svg';
 import Department from '../icons/department.svg';
 import Post from '../icons/post.svg';
 import DotIcon from '../icons/dot.svg';
+import StepsFinsh from '../icons/steps-finsh.svg';
 import Invisible from '../icons/invisible.svg';
 import DateIcon from '../icons/date.svg';
 import InvisibleIcon from '../icons/invisible.svg';
+import DownIcon from '../icons/down.svg';
+import RightDoubleArrowIcon from '../icons/right-double-arrow.svg';
+import NewEdit from '../icons/new-edit.svg';
+import RotateLeft from '../icons/rotate-left.svg';
+import RotateRight from '../icons/rotate-right.svg';
+import ScaleOrigin from '../icons/scale-origin.svg';
 
 // 兼容原来的用法，后续不直接试用。
 
@@ -142,6 +150,7 @@ registerIcon('right-arrow', RightArrowIcon);
 registerIcon('prev', LeftArrowIcon);
 registerIcon('next', RightArrowIcon);
 registerIcon('check', CheckIcon);
+registerIcon('check-circle', CheckCircle);
 registerIcon('plus', PlusIcon);
 registerIcon('add', PlusIcon);
 registerIcon('minus', MinusIcon);
@@ -220,9 +229,16 @@ registerIcon('edit', EditIcon);
 registerIcon('desk-empty', DeskEmptyIcon);
 registerIcon('invisible', Invisible);
 registerIcon('plus-fine', PlusFineIcon);
+registerIcon('steps-finsh', StepsFinsh);
 registerIcon('date', DateIcon);
 registerIcon('remove', RemoveIcon);
 registerIcon('invisible', InvisibleIcon);
+registerIcon('down', DownIcon);
+registerIcon('right-double-arrow', RightDoubleArrowIcon);
+registerIcon('new-edit', NewEdit);
+registerIcon('rotate-left', RotateLeft);
+registerIcon('rotate-right', RotateRight);
+registerIcon('scale-origin', ScaleOrigin);
 
 export function Icon({
   icon,
@@ -266,5 +282,6 @@ export {
   UserRemove,
   Role,
   Department,
-  Post
+  Post,
+  RightDoubleArrowIcon
 };
