@@ -46,7 +46,7 @@ register('zh-CN', {
   'CRUD.paginationGoText': '前往',
   'CRUD.paginationPageText': '页',
   'PaginationWrapper.placeholder': '请配置内容',
-  'Pagination.select': '{{count}} 条/页',
+  'Pagination.select': '{{count}}条/页',
   'Pagination.goto': '跳转至',
   'Pagination.go': 'GO',
   'Pagination.totalPage': '共 {{lastPage}} 页',
@@ -282,6 +282,11 @@ register('zh-CN', {
   'Year.placeholder': '请选择年',
   'reload': '刷新',
   'rotate': '旋转',
+  'rotate.left': '向左旋转',
+  'rotate.right': '向右旋转',
+  'zoomIn': '放大',
+  'zoomOut': '缩小',
+  'scale.origin': '原始尺寸',
   'Editor.fullscreen': '全屏',
   'Editor.exitFullscreen': '退出全屏',
   'Condition.not': '非',
