@@ -323,8 +323,7 @@ test('Renderer:table combineNum', () => {
             {
               type: 'table',
               source: '$rows',
-              className: 'm-b-none',
-              combineNum: 3,
+              combineNum: 1,
               columnsTogglable: false,
               columns: [
                 {
@@ -949,7 +948,7 @@ describe('Renderer:table selectable & itemCheckableOn', () => {
   };
 
   test('radio style', async () => {
-    const {container, debug} = render(amisRender(schema, {}, makeEnv({})));
+    const {container} = render(amisRender(schema, {}, makeEnv({})));
     await waitFor(() => {
       expect(container.querySelector('[type=radio]')).toBeInTheDocument();
     });
