@@ -606,8 +606,26 @@ export default class Table extends React.Component<TableProps, object> {
   }
 
   componentDidMount() {
+    const {autoFillHeight, classPrefix: ns} = this.props;
+    const currentNode = findDOMNode(this) as HTMLElement;
     let parent: HTMLElement | Window | null = getScrollParent(
-      findDOMNode(this) as HTMLElement
+      currentNode,
+      parent => {
+        const parentStyle = getComputedStyle(parent);
+        const currentStyle = getComputedStyle(currentNode);
+
+        const parentHeight =
+          parseInt(parentStyle.getPropertyValue('height')) -
+          parseInt(parentStyle.getPropertyValue('padding-top')) -
+          parseInt(parentStyle.getPropertyValue('padding-bottom'));
+        const currentHeight =
+          parseInt(currentStyle.getPropertyValue('height')) +
+          parseInt(currentStyle.getPropertyValue('margin-bottom')) +
+          parseInt(currentStyle.getPropertyValue('margin-top'));
+
+        // 具备 overflow-*:auto 的父元素的高度小于当前元素
+        return parentHeight > 0 && parentHeight < currentHeight;
+      }
     );
 
     if (!parent || parent === document.body) {
@@ -624,6 +642,11 @@ export default class Table extends React.Component<TableProps, object> {
 
     this.affixDetect();
     parent.addEventListener('scroll', this.affixDetect);
+    if (autoFillHeight) {
+      document
+        .querySelector<HTMLElement>(`.${ns}Table-content`)!
+        .addEventListener('scroll', this.affixDetect);
+    }
     window.addEventListener('resize', this.affixDetect);
     this.updateAutoFillHeight();
     window.addEventListener('resize', this.updateAutoFillHeight);
@@ -798,10 +821,15 @@ export default class Table extends React.Component<TableProps, object> {
   }
 
   componentWillUnmount() {
-    const {formItem} = this.props;
+    const {formItem, autoFillHeight, classPrefix: ns} = this.props;
 
     const parent = this.parentNode;
     parent && parent.removeEventListener('scroll', this.affixDetect);
+    if (autoFillHeight) {
+      document
+        .querySelector<HTMLElement>(`.${ns}Table-content`)!
+        .addEventListener('scroll', this.affixDetect);
+    }
     window.removeEventListener('resize', this.affixDetect);
     window.removeEventListener('resize', this.updateAutoFillHeight);
     (this.updateTableInfoLazy as any).cancel();
@@ -1042,35 +1070,81 @@ export default class Table extends React.Component<TableProps, object> {
 
     return store.selectedRows.map(item => item.data);
   }
-
+  /**
+   * 滚动重新定位Table-fixedTop
+   */
   affixDetect() {
-    if (!this.props.affixHeader || !this.table || this.props.autoFillHeight) {
+    const {
+      affixHeader,
+      classPrefix: ns,
+      affixOffsetTop,
+      env,
+      autoFillHeight
+    } = this.props;
+
+    if (!(affixHeader || autoFillHeight) || !this.table) {
+      // if (!this.props.affixHeader || !this.table || this.props.autoFillHeight) {
       return;
     }
 
-    const ns = this.props.classPrefix;
     const dom = findDOMNode(this) as HTMLElement;
-    const clip = (this.table as HTMLElement).getBoundingClientRect();
-    const offsetY =
-      this.props.affixOffsetTop ?? this.props.env.affixOffsetTop ?? 0;
-    const headingHeight =
-      dom.querySelector(`.${ns}Table-heading`)?.getBoundingClientRect()
-        .height || 0;
     const headerHeight =
       dom.querySelector(`.${ns}Table-headToolbar`)?.getBoundingClientRect()
         .height || 0;
-
-    const affixed =
-      clip.top - headerHeight - headingHeight < offsetY &&
-      clip.top + clip.height - 40 > offsetY;
     const affixedDom = dom.querySelector(`.${ns}Table-fixedTop`) as HTMLElement;
+    const affixedDomHeight =
+      getComputedStyle(affixedDom).getPropertyValue('height');
 
-    affixedDom.style.cssText += `top: ${offsetY}px;width: ${
-      (this.table.parentNode as HTMLElement).offsetWidth
-    }px`;
-    affixed
-      ? affixedDom.classList.add('in')
-      : affixedDom.classList.remove('in');
+    if (autoFillHeight) {
+      //! 解决 sticky 不兼容的问题
+      const tableContentDom = document.querySelector<HTMLElement>(
+        `.${ns}Table-content`
+      )!;
+      const clip = tableContentDom.getBoundingClientRect();
+
+      const tableHeaderClip = affixedDom
+        .querySelector(`.${ns}Table-wrapper`)!
+        .getBoundingClientRect();
+      const offsetHeight =
+        affixedDom.getBoundingClientRect().height - tableHeaderClip.height;
+      const offsetY = clip.top - offsetHeight;
+
+      affixedDom.style.cssText += `top: ${offsetY}px;width: ${clip.width}px`;
+
+      affixedDom.classList.add('in');
+    } else {
+      // 配置了 affixHeader
+      const clip = (this.table as HTMLElement).getBoundingClientRect();
+      const offsetY = affixOffsetTop ?? env.affixOffsetTop ?? 0;
+      const headingHeight =
+        dom.querySelector(`.${ns}Table-heading`)?.getBoundingClientRect()
+          .height || 0;
+      const affixedShadowDom = dom.querySelector(
+        `.${ns}Table-fixedTop-shadow`
+      ) as HTMLElement;
+      const affixed =
+        clip.top - headerHeight - headingHeight < offsetY &&
+        clip.top + clip.height - 40 > offsetY;
+
+      affixedDom.style.cssText += `top: ${offsetY}px;width: ${
+        (this.table.parentNode as HTMLElement).offsetWidth
+      }px`;
+      affixedShadowDom.style.cssText += `top: ${affixedDomHeight};width: ${
+        (this.table.parentNode as HTMLElement).offsetWidth
+      }px`;
+
+      const preStatus = affixedDom.classList.contains('in');
+
+      affixed
+        ? affixedDom.classList.add('in')
+        : affixedDom.classList.remove('in');
+
+      // 出现affixHeader时重新计算列表宽度
+      if (!preStatus && affixed) {
+        this.updateTableInfoLazy();
+      }
+    }
+
     // store.markHeaderAffix(clip.top < offsetY && (clip.top + clip.height - 40) > offsetY);
   }
 
@@ -2105,70 +2179,73 @@ export default class Table extends React.Component<TableProps, object> {
     const columnsGroup = store.columnGroup;
 
     return affixHeader ? (
-      <div
-        className={cx('Table-fixedTop', {
-          'is-fakeHide': hideHeader
-        })}
-      >
-        {this.renderHeader(false)}
-        {this.renderHeading()}
-        <div className={cx('Table-fixedLeft')}>
-          {store.leftFixedColumns.length
-            ? this.renderFixedColumns(
-                store.rows,
-                store.leftFixedColumns,
-                true,
-                tableClassName
-              )
-            : null}
-        </div>
-        <div className={cx('Table-fixedRight')}>
-          {store.rightFixedColumns.length
-            ? this.renderFixedColumns(
-                store.rows,
-                store.rightFixedColumns,
-                true,
-                tableClassName
-              )
-            : null}
-        </div>
-        <div className={cx('Table-wrapper')}>
-          <table ref={this.affixedTableRef} className={tableClassName}>
-            <colgroup>
-              {store.filteredColumns.map(column => (
-                <col key={column.index} data-index={column.index} />
-              ))}
-            </colgroup>
-            <thead>
-              {columnsGroup.length ? (
+      <>
+        <div
+          className={cx('Table-fixedTop', {
+            'is-fakeHide': hideHeader
+          })}
+        >
+          {this.renderHeader(false)}
+          {this.renderHeading()}
+          <div className={cx('Table-fixedLeft')}>
+            {store.leftFixedColumns.length
+              ? this.renderFixedColumns(
+                  store.rows,
+                  store.leftFixedColumns,
+                  true,
+                  tableClassName
+                )
+              : null}
+          </div>
+          <div className={cx('Table-fixedRight')}>
+            {store.rightFixedColumns.length
+              ? this.renderFixedColumns(
+                  store.rows,
+                  store.rightFixedColumns,
+                  true,
+                  tableClassName
+                )
+              : null}
+          </div>
+          <div className={cx('Table-wrapper')}>
+            <table ref={this.affixedTableRef} className={tableClassName}>
+              <colgroup>
+                {store.filteredColumns.map(column => (
+                  <col key={column.index} data-index={column.index} />
+                ))}
+              </colgroup>
+              <thead>
+                {columnsGroup.length ? (
+                  <tr>
+                    {columnsGroup.map((item, index) => (
+                      <th
+                        key={index}
+                        data-index={item.index}
+                        colSpan={item.colSpan}
+                        rowSpan={item.rowSpan}
+                      >
+                        {item.label ? render('tpl', item.label) : null}
+                      </th>
+                    ))}
+                  </tr>
+                ) : null}
                 <tr>
-                  {columnsGroup.map((item, index) => (
-                    <th
-                      key={index}
-                      data-index={item.index}
-                      colSpan={item.colSpan}
-                      rowSpan={item.rowSpan}
-                    >
-                      {item.label ? render('tpl', item.label) : null}
-                    </th>
-                  ))}
+                  {store.filteredColumns.map(column =>
+                    columnsGroup.find(group => ~group.has.indexOf(column))
+                      ?.rowSpan === 2
+                      ? null
+                      : this.renderHeadCell(column, {
+                          'key': column.index,
+                          'data-index': column.index
+                        })
+                  )}
                 </tr>
-              ) : null}
-              <tr>
-                {store.filteredColumns.map(column =>
-                  columnsGroup.find(group => ~group.has.indexOf(column))
-                    ?.rowSpan === 2
-                    ? null
-                    : this.renderHeadCell(column, {
-                        'key': column.index,
-                        'data-index': column.index
-                      })
-                )}
-              </tr>
-            </thead>
-          </table>
+              </thead>
+            </table>
+          </div>
         </div>
-      </div>
+        <div className={cx('Table-fixedTop-shadow')}></div>
+      </>
     ) : null;
   }
 
@@ -2198,11 +2275,9 @@ export default class Table extends React.Component<TableProps, object> {
     const columnsGroup = store.columnGroup;
     return (
       <table
-        className={cx(
-          'Table-table',
-          store.combineNum > 0 ? 'Table-table--withCombine' : '',
-          tableClassName
-        )}
+        className={cx('Table-table', tableClassName, {
+          'Table-table--withCombine': store.combineNum > 0
+        })}
       >
         <thead>
           {columnsGroup.length ? (
@@ -2249,10 +2324,9 @@ export default class Table extends React.Component<TableProps, object> {
           </tbody>
         ) : (
           <TableBody
-            tableClassName={cx(
-              store.combineNum > 0 ? 'Table-table--withCombine' : '',
-              tableClassName
-            )}
+            tableClassName={cx(tableClassName, {
+              'Table-table--withCombine': store.combineNum > 0
+            })}
             itemAction={itemAction}
             classnames={cx}
             render={render}
@@ -2702,8 +2776,10 @@ export default class Table extends React.Component<TableProps, object> {
       <>
         <TableContent
           tableClassName={cx(
-            store.combineNum > 0 ? 'Table-table--withCombine' : '',
-            {'Table-table--checkOnItemClick': checkOnItemClick},
+            {
+              'Table-table--checkOnItemClick': checkOnItemClick,
+              'Table-table--withCombine': store.combineNum > 0
+            },
             tableClassName
           )}
           className={tableContentClassName}
@@ -2795,11 +2871,9 @@ export default class Table extends React.Component<TableProps, object> {
     const heading = this.renderHeading();
     const header = this.renderHeader();
     const footer = this.renderFooter();
-    const tableClassName = cx(
-      'Table-table',
-      store.combineNum > 0 ? 'Table-table--withCombine' : '',
-      this.props.tableClassName
-    );
+    const tableClassName = cx('Table-table', this.props.tableClassName, {
+      'Table-table--withCombine': store.combineNum > 0
+    });
 
     return (
       <div
